split sdk per entity
